requirements.txt
Django
django-heroku
djangorestframework
gunicorn
Pillow
scikit-image
whitenoise
numpy
scipy
tensorflow-cpu
